🎨 Moved hook to hook folder
