+++ src/app/utils/user.tsx
@@ -1,5 +1,3 @@
-import { useEffect, useState } from 'react';
-
 export type Profile = {
   username: string | null;
   displayName: string | null;
@@ -31,28 +29,4 @@ export function getCurrentUser(): Promise<Profile | null> {
       reject(error);
     }
   });
-}
-
-export function useCurrentUser(): [Profile | null, string | null] {
-  const [currentUser, setProfile] = useState<Profile | null>(null);
-  const [profileError, setProfileError] = useState<string | null>(null);
-
-  useEffect(() => {
-    async function loadUser() {
-      try {
-        const currentUser = await getCurrentUser();
-        setProfile(currentUser);
-      } catch (error) {
-        console.error(error);
-        if (error instanceof Error) {
-          setProfileError(error.message);
-        }
-      }
-    }
-    loadUser();
-    overwolf.profile.onLoginStateChanged.addListener(loadUser);
-    return () => overwolf.profile.onLoginStateChanged.removeListener(loadUser);
-  }, []);
-
-  return [currentUser, profileError];
 }
